Add feature and redesign number system
Add calculation of percentages from inputted decimals.

Redesign double based number system to BigDecimal number system.

The double based number system was causing rounding errors in some
calculations. The newly implemented BigDecimal based system is more
precise and controllable.

+++ src/main/java/org/calculator/processing/OperationStrategy.java
@@ -2,6 +2,8 @@ package org.calculator.processing;
 
 import org.calculator.common.Request;
 
+import java.math.BigDecimal;
+
 interface OperationStrategy {
-	double execute(Request request);
+	BigDecimal execute(Request request);
 }
