+++ src/main/java/org/calculator/processing/ProcessorContext.java
@@ -2,12 +2,14 @@ package org.calculator.processing;
 
 import org.calculator.common.Request;
 
+import java.math.BigDecimal;
+
 class ProcessorContext {
 	private OperationStrategy strategy;
 	public void setStrategy(OperationStrategy strategy){
 		this.strategy = strategy;
 	}
-	public double executeStrategy(Request request){
+	public BigDecimal executeStrategy(Request request){
 		return strategy.execute(request);
 	}
 }
