+++ src/main/java/org/calculator/request/RequestBuilder.java
@@ -3,6 +3,8 @@ package org.calculator.request;
 import org.calculator.common.Operations;
 import org.calculator.common.Request;
 
+import java.math.BigDecimal;
+
 class RequestBuilder implements Builder{
 	private Request request;
 
@@ -14,7 +16,7 @@ class RequestBuilder implements Builder{
 		return request;
 	}
 
-	public void buildSection(Operations section, double[] values) {
+	public void buildSection(Operations section, BigDecimal[] values) {
 		request.setSection(section, values);
 	}
 }
